chore: refactor physical join
